+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Bug fixes:
 * Fixed incorrect comparisons between bignums and doubles. 
 * Prevented some internal uses of `Kernel#caller_locations` to be overridden by user code (#1934). 
 * Fixed an issue caused by recursing inlining within `Regexp#quote` (#1927).
+* Updated `Kernel.Float()` to return given string in error message (#1945).
 
 Compatibility: